gem: extract test helpers
